+++ noodles-cram/src/io/writer.rs
@@ -11,7 +11,7 @@ pub(crate) mod record;
 use std::io::{self, Write};
 
 use noodles_fasta as fasta;
-use noodles_sam as sam;
+use noodles_sam::{self as sam, alignment::io::Write as _};
 
 pub use self::builder::Builder;
 use self::{
@@ -228,13 +228,21 @@ where
     /// let header = sam::Header::default();
     /// writer.write_header(&header)?;
     ///
-    /// let record = sam::Record::default();
-    /// writer.write_alignment_record(&header, &record)?;
+    /// let record = cram::Record::default();
+    /// writer.write_record(&header, &record)?;
     ///
     /// writer.try_finish(&header)?;
     /// # Ok::<(), io::Error>(())
     /// ```
-    pub fn write_record(&mut self, header: &sam::Header, record: Record) -> io::Result<()> {
+    pub fn write_record(
+        &mut self,
+        header: &sam::Header,
+        record: &crate::Record<'_>,
+    ) -> io::Result<()> {
+        self.write_alignment_record(header, record)
+    }
+
+    fn add_record(&mut self, header: &sam::Header, record: Record) -> io::Result<()> {
         self.records.push(record);
 
         if self.records.len() >= self.records.capacity() {
@@ -277,8 +285,8 @@ where
         header: &sam::Header,
         record: &dyn sam::alignment::Record,
     ) -> io::Result<()> {
-        let r = Record::try_from_alignment_record(header, record)?;
-        self.write_record(header, r)
+        let record = Record::try_from_alignment_record(header, record)?;
+        self.add_record(header, record)
     }
 
     fn finish(&mut self, header: &sam::Header) -> io::Result<()> {